requirements.txt
Flask==2.0.2
Flask_Bootstrap==3.3.7.1
Flask_MySQLdb==0.2.0
PyYAML==6.0
gunicorn
